refactor: move edge websocket protocol to astra-server-types
