+++ rust/crates/astra-server-types/src/edge_ws_protocol.rs
@@ -1,0 +1,199 @@
+//! Edge WebSocket protocol types for remote tool execution.
+//!
+//! Defines the bidirectional message protocol between the Astra server and
+//! remote edge agents. Edge agents connect via `GET /edge/ws`, authenticate,
+//! and then receive tool execution requests and return results.
+//!
+//! ## Protocol
+//!
+//! **Edge → Server** (JSON text frames):
+//! ```text
+//! {"type": "edge_auth", "token": "Bearer ...", "edge_agent_id": "...", "hostname": "...", "workspace_dir": "..."}
+//! {"type": "edge_tool_result", "request_id": "...", "output": "...", "is_error": false}
+//! {"type": "edge_ping"}
+//! ```
+//!
+//! **Server → Edge** (JSON text frames):
+//! ```text
+//! {"type": "edge_auth_ok", "user_id": "..."}
+//! {"type": "edge_auth_error", "message": "..."}
+//! {"type": "edge_tool_request", "request_id": "...", "tool": "...", "args": {...}}
+//! {"type": "edge_pong"}
+//! {"type": "edge_closing", "reason": "..."}
+//! ```
+
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+
+/// Timeout for tool execution on the edge agent.
+pub const EDGE_TOOL_TIMEOUT_SECS: u64 = 300; // 5 minutes
+
+/// Timeout for the initial auth message after WebSocket upgrade.
+pub const EDGE_AUTH_TIMEOUT_SECS: u64 = 30;
+
+/// Heartbeat interval for edge keep-alive.
+pub const EDGE_HEARTBEAT_INTERVAL_SECS: u64 = 30;
+
+/// Messages sent from edge agent to server.
+#[derive(Deserialize, Debug, Clone)]
+#[serde(tag = "type")]
+pub enum EdgeClientMessage {
+    /// Authenticate the edge agent (must be first message).
+    #[serde(rename = "edge_auth")]
+    Auth {
+        token: String,
+        edge_agent_id: String,
+        #[serde(default)]
+        hostname: Option<String>,
+        #[serde(default)]
+        workspace_dir: Option<String>,
+        #[serde(default)]
+        capabilities: Option<Vec<String>>,
+    },
+
+    /// Tool execution result from the edge.
+    #[serde(rename = "edge_tool_result")]
+    ToolResult {
+        request_id: String,
+        output: String,
+        #[serde(default)]
+        is_error: bool,
+        #[serde(default)]
+        duration_ms: Option<u64>,
+    },
+
+    /// Edge heartbeat.
+    #[serde(rename = "edge_ping")]
+    Ping,
+}
+
+/// Messages sent from server to edge agent.
+#[derive(Serialize, Debug, Clone)]
+#[serde(tag = "type")]
+pub enum EdgeServerMessage {
+    /// Authentication succeeded.
+    #[serde(rename = "edge_auth_ok")]
+    AuthOk { user_id: String },
+
+    /// Authentication failed.
+    #[serde(rename = "edge_auth_error")]
+    AuthError { message: String },
+
+    /// Request the edge to execute a tool.
+    #[serde(rename = "edge_tool_request")]
+    ToolRequest {
+        request_id: String,
+        tool: String,
+        args: Value,
+        /// Maximum execution time in seconds.
+        timeout_secs: u64,
+    },
+
+    /// Server heartbeat response.
+    #[serde(rename = "edge_pong")]
+    Pong,
+
+    /// Server is closing the connection.
+    #[serde(rename = "edge_closing")]
+    Closing { reason: String },
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use serde_json::json;
+
+    #[test]
+    fn edge_auth_deserializes() {
+        let msg: EdgeClientMessage = serde_json::from_value(json!({
+            "type": "edge_auth",
+            "token": "Bearer abc",
+            "edge_agent_id": "my-edge",
+            "hostname": "laptop",
+            "workspace_dir": "/home/user/project"
+        }))
+        .unwrap();
+        match msg {
+            EdgeClientMessage::Auth {
+                token,
+                edge_agent_id,
+                hostname,
+                ..
+            } => {
+                assert_eq!(token, "Bearer abc");
+                assert_eq!(edge_agent_id, "my-edge");
+                assert_eq!(hostname.as_deref(), Some("laptop"));
+            }
+            _ => panic!("expected Auth"),
+        }
+    }
+
+    #[test]
+    fn edge_tool_result_deserializes() {
+        let msg: EdgeClientMessage = serde_json::from_value(json!({
+            "type": "edge_tool_result",
+            "request_id": "req-123",
+            "output": "file contents here",
+            "is_error": false,
+            "duration_ms": 42
+        }))
+        .unwrap();
+        match msg {
+            EdgeClientMessage::ToolResult {
+                request_id,
+                output,
+                is_error,
+                duration_ms,
+            } => {
+                assert_eq!(request_id, "req-123");
+                assert_eq!(output, "file contents here");
+                assert!(!is_error);
+                assert_eq!(duration_ms, Some(42));
+            }
+            _ => panic!("expected ToolResult"),
+        }
+    }
+
+    #[test]
+    fn edge_tool_request_serializes() {
+        let msg = EdgeServerMessage::ToolRequest {
+            request_id: "req-456".into(),
+            tool: "bash".into(),
+            args: json!({"command": "echo hello"}),
+            timeout_secs: 120,
+        };
+        let v = serde_json::to_value(&msg).unwrap();
+        assert_eq!(v["type"], "edge_tool_request");
+        assert_eq!(v["tool"], "bash");
+        assert_eq!(v["timeout_secs"], 120);
+    }
+
+    #[test]
+    fn edge_ping_deserializes() {
+        let msg: EdgeClientMessage = serde_json::from_value(json!({"type": "edge_ping"})).unwrap();
+        assert!(matches!(msg, EdgeClientMessage::Ping));
+    }
+
+    #[test]
+    fn edge_pong_serializes() {
+        let msg = EdgeServerMessage::Pong;
+        let v = serde_json::to_value(&msg).unwrap();
+        assert_eq!(v["type"], "edge_pong");
+    }
+
+    #[test]
+    fn edge_auth_ok_serializes() {
+        let msg = EdgeServerMessage::AuthOk {
+            user_id: "u-123".into(),
+        };
+        let v = serde_json::to_value(&msg).unwrap();
+        assert_eq!(v["type"], "edge_auth_ok");
+        assert_eq!(v["user_id"], "u-123");
+    }
+
+    const _: () = {
+        assert!(EDGE_TOOL_TIMEOUT_SECS >= 60);
+        assert!(EDGE_AUTH_TIMEOUT_SECS >= 10);
+        assert!(EDGE_HEARTBEAT_INTERVAL_SECS >= 10);
+    };
+}
