+++ rust/crates/astra-server-types/src/lib.rs
@@ -1,3 +1,5 @@
+mod edge_ws_protocol;
+
 use astra_services::auth::SessionActivityRecord;
 use astra_services::{
     AdminAuditRecord, AdminFeedbackStatsRecord, AdminInitRecord, AdminTokenRecord,
@@ -6,6 +8,11 @@ use astra_services::{
     SessionRecord,
 };
 use serde::{Deserialize, Serialize};
+
+pub use edge_ws_protocol::{
+    EDGE_AUTH_TIMEOUT_SECS, EDGE_HEARTBEAT_INTERVAL_SECS, EDGE_TOOL_TIMEOUT_SECS,
+    EdgeClientMessage, EdgeServerMessage,
+};
 
 #[derive(Serialize, PartialEq, Eq)]
 pub struct RootResponse {
